Latest.

+++ update-deps.sh
@@ -12,30 +12,30 @@ echo
 echo "compress node_modules"
 tar cjf node_modules.tar.bz2 node_modules/
 
-echo "upload to seperate github repo"
-CACHE_DIR=node_modules_cache
-CACHE_REPO=git@github.com:saavuio/s_nuxt_2nd_cache.git
-VERSION=v1
+if [ ! -z "$PUSH_TO_REMOTE" ]; then
+  echo "upload to seperate github repo"
+  CACHE_DIR=node_modules_cache
+  CACHE_REPO=git@github.com:saavuio/s_nuxt_2nd_cache.git
+  VERSION=v1
 
-if [ -d $CACHE_DIR ]; then
-  rm -rf $CACHE_DIR
+  if [ -d $CACHE_DIR ]; then
+    rm -rf $CACHE_DIR
+  fi
+
+  mkdir $CACHE_DIR
+  mv node_modules.tar.bz2 $CACHE_DIR
+
+  if [ ! -z "$1" ]; then
+    TARGET_BRANCH=main-repo-sha-$1
+  else
+    TARGET_BRANCH=$VERSION
+  fi
+
+  cd $CACHE_DIR
+  git init
+  git remote add origin $CACHE_REPO
+  git checkout -b $TARGET_BRANCH
+  git add node_modules.tar.bz2
+  git commit -m "Latest."
+  git push -u -f origin $TARGET_BRANCH
 fi
-
-mkdir $CACHE_DIR
-mv node_modules.tar.bz2 $CACHE_DIR
-
-# UNCOMMENT TO PUSH TO GITHUB
-
-if [ ! -z "$1" ]; then
-  TARGET_BRANCH=main-repo-sha-$1
-else
-  TARGET_BRANCH=$VERSION
-fi
-
-cd $CACHE_DIR
-git init
-git remote add origin $CACHE_REPO
-git checkout -b $TARGET_BRANCH
-git add node_modules.tar.bz2
-git commit -m "Latest."
-git push -u -f origin $TARGET_BRANCH
